+++ CHANGELOG.md
@@ -6,6 +6,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ### [Unreleased]
 ### Changed
 - Added framework.events.received to NewSubscriptionManagerDelegate to count correctly
+- Added new method `cleanViewStoreErrorTables()` in DatabaseCleaner test utility class to clean the new error tables in the viewstore
 
 # [17.103.0-M23] - 2025-07-07
 ### Added